refactor: cleaned-up ContactsTableViewCell to follow SOLID principles
